Add api_key file to .gitignore
Since it is not good practice to push api_keys, we are following
that as a standard even though this is a private_repo.

Please obtain the file 'api_key.xml' from me personally and include it
inside 'Values'.

Remember to pass api_key.xml in the submitted repo upon completion of the
Team Project for class.